Refactoring Exercise part 2
